+++ nvim/lua/custom/plugins/wikivim.lua
@@ -1,3 +1,0 @@
-return {
-  'lervag/wiki.vim', -- Vimwiki setup for zettelkasten
-}