+++ nvim/lua/plugins/colour.lua
@@ -1,0 +1,4 @@
+return {
+  { "catppuccin/nvim", name = "catppuccin" },
+  { "LazyVim/LazyVim", opts = { colorscheme = "catppuccin-frappe" } },
+}