+++ nvim/lua/plugins/language.lua
@@ -1,0 +1,55 @@
+return {
+  {
+    "nvim-treesitter/nvim-treesitter",
+    opts = {
+      ensure_installed = {
+        "bash",
+        "fish",
+        "html",
+        "javascript",
+        "json",
+        "lua",
+        "markdown",
+        "markdown_inline",
+        "python",
+        "query",
+        "regex",
+        "vim",
+        "yaml",
+      },
+    },
+  },
+  {
+    "stevearc/conform.nvim",
+    opts = {
+      formatters_by_ft = {
+        quarto = { "injected" },
+        sql = { "sqlfmt" },
+      },
+      -- See:
+      -- https://github.com/jmbuhr/quarto-nvim-kickstarter/blob/382b050e13eada7180ad048842386be37e820660/lua/plugins/editing.lua#L29-L81
+      injected = {
+        -- Set the options field
+        options = {
+          -- Set to true to ignore errors
+          ignore_errors = false,
+          -- Map of treesitter language to file extension
+          -- A temporary file name with this extension will be generated during formatting
+          -- because some formatters care about the filename.
+          lang_to_ext = {
+            bash = "sh",
+            javascript = "js",
+            latex = "tex",
+            markdown = "md",
+            python = "py",
+            r = "r",
+            typescript = "ts",
+          },
+          -- Map of treesitter language to formatters to use
+          -- (defaults to the value from formatters_by_ft)
+          lang_to_formatters = {},
+        },
+      }
+    },
+  },
+}